+++ src/LuminaPath.Core/Dtos/Media/MyGameDto.cs
@@ -13,7 +13,7 @@ namespace LuminaPath.Core.Dtos
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
         public GameStatus Status { get; set; }
-        public int? TimeSpend { get; set; }
+        public double? TimeSpend { get; set; }
         public int GameId { get; set; }
         public string? PersonalNotes { get; set; }
         public GamesNoIncludeDto? Game { get; set; }
